feat(shokunin): decoupling write_element
